Refine global search capture ownership
